优化通信机制

+++ SamSoarII/Core/Communication/Command/Download/PLCMessage.cs
@@ -31,7 +31,7 @@ namespace SamSoarII.Core.Communication
         public int Ver3;
 
         //站号
-        public int StationNumber;
+        public byte StationNumber;
 
         //PLC类型
         public object PLCType;
